refactor: ImageService와 ImageUploader의 위치 이동
